requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
colorama==0.3.9
grappa==0.1.9
more-itertools==4.3.0
Pillow==5.2.0
pluggy==0.7.1
py==1.5.4
pytest==3.7.1
six==1.11.0
